# painter-2026-unuseable-ai/packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Painter 2026 API
  version: 1.0.0
  description: Contract-first APIs for painter-2026 services.
servers:
  - url: /api/v1
paths:
  /identity/auth/login:
    post:
      summary: User login
      operationId: identityLogin
      tags: [identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginData'
...
  /identity/follows:
    get:
      summary: List followings by user
      operationId: listFollowings
      tags: [identity]
      parameters:
        - in: query
          name: userId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Followings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseFollowListData'
    post:
      summary: Follow user
      operationId: createFollow
      tags: [identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowRequest'
      responses:
        '200':
          description: Follow action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseActionData'
...
components:
  schemas:
    ApiMeta:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        traceId: { type: string }
      required: [code, message, traceId]
    ApiResponseLoginData:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/LoginData'
...
    ApiResponseActionData:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data:
              type: object
              properties:
                success: { type: boolean }
              required: [success]
    ApiResponseFollowListData:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/FollowItem'
              required: [items]
...
    LoginRequest:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
      required: [username, password]
    LoginData:
      type: object
      properties:
        userId: { type: string }
        accessToken: { type: string }
        expiresAt: { type: string, format: date-time }
      required: [userId, accessToken, expiresAt]
...
    FollowRequest:
      type: object
      properties:
        targetUserId: { type: string }
      required: [targetUserId]
    FollowItem:
      type: object
      properties:
        userId: { type: string }
        nickname: { type: string }
      required: [userId, nickname]
...